+++ oalg-base/CHANGELOG.md
@@ -94,11 +94,12 @@ has been defined by EmbeddingFrom.
 - class Filterable in OAlg.Data.Filterable.
 - structure EntOrd for ordered entities.
 - module OAlg.Data.Logical
+- module OAlg.Category.Map
 - module OAlg.Boolean.Definition: class Logical as base class for Booleans, class Erasable.
 - classes PartiallyOrdered, Empty, Full in OAlg.Structure.PartiallyOrdered.Definition
 - class Lattice in OAlg.Structure.Lattice.Definition.
 - module OAlg.Entity.Seqeuence.Set: setPower, setIsEmpty, setIntersection, setDifference, setTakeN
-  setFilter, instance Erasable Set x, OrdMap, EntOrdMap.
+  setFilter, instance Erasable (Set x)
 - module OAlg.Entity.Sequence.Graph: gphTakeN, gphset, setgph, gphUnion, gphIntersection,
 gphSetFilter, gphDifference, isSubGraph, instance Ord (Graph i x), Lattice (Graph a (Set b)),
 instance Filterable (Graph i).